services/api-gateway/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
pyjwt==2.10.1
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
